+++ packages/components-react/projects/nextjs/tests/vrt/specs/components.vrt.ts
@@ -26,7 +26,7 @@ const components = fileNames
   });
 
 test('should have certain amount of components', () => {
-  expect(components.length).toBe(60);
+  expect(components.length).toBe(61);
 });
 
 for (const component of components) {