+++ packages/components-vue/src/examples/InputNumberExampleControlled.vue
@@ -1,0 +1,20 @@
+<script setup lang="ts">
+import { type InputNumberInputEventDetail, PInputNumber, PText } from '@porsche-design-system/components-vue';
+import { computed, ref } from 'vue';
+
+const value = ref<string>('');
+
+const onInput = (e: InputNumberInputEventDetail) => {
+  console.log(e);
+  value.value = (e.target as HTMLInputElement).value;
+};
+
+const debugText = computed(() => {
+  return `Value: ${value.value}`;
+});
+</script>
+
+<template>
+  <PInputNumber :name="'some-name'" :label="'Some Label'" :value="value" @input="onInput" />
+  <PText>{{ debugText }}</PText>
+</template>