+++ packages/components-vue/src/examples/index.ts
@@ -24,6 +24,8 @@ export { default as InlineNotificationExampleActionButton } from './InlineNotifi
 export { default as InlineNotificationExampleEvents } from './InlineNotificationExampleEvents.vue';
 export { default as InputPasswordExample } from './InputPasswordExample.vue';
 export { default as InputPasswordExampleControlled } from './InputPasswordExampleControlled.vue';
+export { default as InputNumberExample } from './InputNumberExample.vue';
+export { default as InputNumberExampleControlled } from './InputNumberExampleControlled.vue';
 export { default as LinkTileExampleHyphens } from './LinkTileExampleHyphens.vue';
 export { default as LinkTileProductExample } from './LinkTileProductExample.vue';
 export { default as ModalExampleAccessibility } from './ModalExample.vue';